+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.0.3
 - Added multi-baseline ablation game. This game computes ablation paths with respect to multiple baseline configurations and aggregates values for different paths via mean, min, max or variance.
+- Added waterfall plots to the HyperSHAP interface.
 
 # v0.0.2
 - Added parallelization for faster analysis